tgstation fork update

+++ code/modules/antagonists/ert/ert.dm
@@ -8,15 +8,17 @@
 	var/datum/team/ert/ert_team
 	var/leader = FALSE
 	var/datum/outfit/outfit = /datum/outfit/centcom/ert/security
+	var/datum/outfit/plasmaman_outfit = /datum/outfit/plasmaman/centcom_official
 	var/role = "Security Officer"
 	var/list/name_source
 	var/random_names = TRUE
 	var/rip_and_tear = FALSE
 	var/equip_ert = TRUE
 	var/forge_objectives_for_ert = TRUE
+	can_elimination_hijack = ELIMINATION_PREVENT
 	show_in_antagpanel = FALSE
+	show_to_ghosts = TRUE
 	antag_moodlet = /datum/mood_event/focused
-	can_hijack = HIJACK_PREVENT
 
 /datum/antagonist/ert/on_gain()
 	if(random_names)
@@ -37,15 +39,32 @@
 /datum/antagonist/ert/proc/update_name()
 	owner.current.fully_replace_character_name(owner.current.real_name,"[role] [pick(name_source)]")
 
-/datum/antagonist/ert/deathsquad/New()
-	. = ..()
-	name_source = GLOB.commando_names
+/datum/antagonist/ert/official
+	name = "CentCom Official"
+	show_name_in_check_antagonists = TRUE
+	var/datum/objective/mission
+	role = "Inspector"
+	random_names = FALSE
+	outfit = /datum/outfit/centcom/centcom_official
 
-/datum/antagonist/ert/deathsquad/apply_innate_effects(mob/living/mob_override)
-	ADD_TRAIT(owner, TRAIT_DISK_VERIFIER, DEATHSQUAD_TRAIT)
+/datum/antagonist/ert/official/greet()
+	to_chat(owner, "<B><font size=3 color=red>You are a CentCom Official.</font></B>")
+	if (ert_team)
+		to_chat(owner, "Central Command is sending you to [station_name()] with the task: [ert_team.mission.explanation_text]")
+	else
+		to_chat(owner, "Central Command is sending you to [station_name()] with the task: [mission.explanation_text]")
 
-/datum/antagonist/ert/deathsquad/remove_innate_effects(mob/living/mob_override)
-	REMOVE_TRAIT(owner, TRAIT_DISK_VERIFIER, DEATHSQUAD_TRAIT)
+/datum/antagonist/ert/official/forge_objectives()
+	if (ert_team)
+		return ..()
+	if(mission)
+		return
+	var/datum/objective/missionobj = new ()
+	missionobj.owner = owner
+	missionobj.explanation_text = "Conduct a routine performance review of [station_name()] and its Captain."
+	missionobj.completed = TRUE
+	mission = missionobj
+	objectives |= mission
 
 /datum/antagonist/ert/security // kinda handled by the base template but here for completion
 
@@ -69,15 +88,34 @@
 /datum/antagonist/ert/commander
 	role = "Commander"
 	outfit = /datum/outfit/centcom/ert/commander
+	plasmaman_outfit = /datum/outfit/plasmaman/centcom_commander
 
 /datum/antagonist/ert/commander/red
 	outfit = /datum/outfit/centcom/ert/commander/alert
 
+/datum/antagonist/ert/janitor
+	role = "Janitor"
+	outfit = /datum/outfit/centcom/ert/janitor
+
+/datum/antagonist/ert/janitor/heavy
+	role = "Heavy Duty Janitor"
+	outfit = /datum/outfit/centcom/ert/janitor/heavy
+
 /datum/antagonist/ert/deathsquad
 	name = "Deathsquad Trooper"
 	outfit = /datum/outfit/centcom/death_commando
+	plasmaman_outfit = /datum/outfit/plasmaman/centcom_commander
 	role = "Trooper"
 	rip_and_tear = TRUE
+
+/datum/antagonist/ert/deathsquad/New()
+	. = ..()
+	name_source = GLOB.commando_names
+
+/datum/antagonist/ert/deathsquad/leader
+	name = "Deathsquad Officer"
+	outfit = /datum/outfit/centcom/death_commando
+	role = "Officer"
 
 /datum/antagonist/ert/medic/inquisitor
 	outfit = /datum/outfit/centcom/ert/medic/inquisitor
@@ -111,33 +149,29 @@
 	. = ..()
 	owner.holy_role = HOLY_ROLE_PRIEST
 
-/datum/antagonist/ert/janitor
-	role = "Janitor"
-	outfit = /datum/outfit/centcom/ert/janitor
-
-/datum/antagonist/ert/janitor/heavy
-	role = "Heavy Duty Janitor"
-	outfit = /datum/outfit/centcom/ert/janitor/heavy
-
-/datum/antagonist/ert/deathsquad/leader
-	name = "Deathsquad Officer"
-	outfit = /datum/outfit/centcom/death_commando
-	role = "Officer"
-
 /datum/antagonist/ert/intern
 	name = "CentCom Intern"
 	outfit = /datum/outfit/centcom/centcom_intern
+	plasmaman_outfit = /datum/outfit/plasmaman/centcom_intern
 	random_names = FALSE
 	role = "Intern"
 
 /datum/antagonist/ert/intern/leader
 	name = "CentCom Head Intern"
 	outfit = /datum/outfit/centcom/centcom_intern/leader
+	random_names = FALSE
 	role = "Head Intern"
+
+/datum/antagonist/ert/intern/unarmed
+	outfit = /datum/outfit/centcom/centcom_intern/unarmed
+
+/datum/antagonist/ert/intern/leader/unarmed
+	outfit = /datum/outfit/centcom/centcom_intern/leader/unarmed
 
 /datum/antagonist/ert/clown
 	role = "Clown"
 	outfit = /datum/outfit/centcom/ert/clown
+	plasmaman_outfit = /datum/outfit/plasmaman/party_comedian
 
 /datum/antagonist/ert/clown/New()
 	. = ..()
@@ -146,14 +180,17 @@
 /datum/antagonist/ert/janitor/party
 	role = "Party Cleaning Service"
 	outfit = /datum/outfit/centcom/ert/janitor/party
+	plasmaman_outfit = /datum/outfit/plasmaman/party_janitor
 
 /datum/antagonist/ert/security/party
 	role = "Party Bouncer"
 	outfit = /datum/outfit/centcom/ert/security/party
+	plasmaman_outfit = /datum/outfit/plasmaman/party_bouncer
 
 /datum/antagonist/ert/engineer/party
 	role = "Party Constructor"
 	outfit = /datum/outfit/centcom/ert/engineer/party
+	plasmaman_outfit = /datum/outfit/plasmaman/party_constructor
 
 /datum/antagonist/ert/clown/party
 	role = "Party Comedian"
@@ -175,6 +212,10 @@
 	var/mob/living/carbon/human/H = owner.current
 	if(!istype(H))
 		return
+	if(isplasmaman(H))
+		H.equipOutfit(plasmaman_outfit)
+		H.internal = H.get_item_for_held_index(2)
+		H.update_internals_hud_icon(1)
 	H.equipOutfit(outfit)
 
 
@@ -190,7 +231,7 @@
 	else
 		missiondesc += " Follow orders given to you by your squad leader."
 	if(!rip_and_tear)
-		missiondesc += "Avoid civilian casualties when possible."
+		missiondesc += " Avoid civilian casualties when possible."
 
 	missiondesc += "<BR><B>Your Mission</B> : [ert_team.mission.explanation_text]"
 	to_chat(owner,missiondesc)
@@ -205,29 +246,48 @@
 	..()
 	var/mob/living/M = mob_override || owner.current
 	add_antag_hud(antag_hud_type, antag_hud_name, M)
+	if(M.hud_used)
+		var/datum/hud/H = M.hud_used
+		var/atom/movable/screen/wanted/giving_wanted_lvl = new /atom/movable/screen/wanted()
+		H.wanted_lvl = giving_wanted_lvl
+		giving_wanted_lvl.hud = H
+		H.infodisplay += giving_wanted_lvl
+		H.mymob.client.screen += giving_wanted_lvl
 
 
 /datum/antagonist/ert/families/remove_innate_effects(mob/living/mob_override)
 	var/mob/living/M = mob_override || owner.current
 	remove_antag_hud(antag_hud_type, M)
+	if(M.hud_used)
+		var/datum/hud/H = M.hud_used
+		H.infodisplay -= H.wanted_lvl
+		QDEL_NULL(H.wanted_lvl)
 	..()
 
 /datum/antagonist/ert/families/greet()
-	to_chat(owner, "<B><font size=3 color=red>You are the [name].</font></B>")
-
+	to_chat(owner, "<B><font size=6 color=red>You are the [name].</font></B>")
+	to_chat(owner, "<B><font size=5 color=red>You are NOT a Nanotrasen Employee. You work for the local government.</font></B>")
+	to_chat(owner, "<B><font size=5 color=red>You are NOT a deathsquad. You are here to help innocents escape violence, criminal activity, and other dangerous things.</font></B>")
 	var/missiondesc = "After an uptick in gang violence on [station_name()], you are responding to emergency calls from the station for immediate SSC Police assistance!\n"
 	missiondesc += "<BR><B>Your Mission</B>:"
-	missiondesc += "<BR> <B>1.</B> Secure the situation and crack down on any gang activity. You can view gangsters with your sunglasses."
-	missiondesc += "<BR> <B>2.</B> There is an undercover police officer on station. Secure him, receive his intel, and extract him safely."
-	missiondesc += "<BR> <B>3.</B> Minimize civilian casualties, but defend yourself and civilians from hostile gangsters."
-	missiondesc += "<BR> <B>3.</B> If Security is found to be violating the rights of citizens, detain them as per your authority as Spinward Stellar Coalition officers."
-	missiondesc += "<BR> <B>4.</B> If the situation demands it, evacuate the station. Otherwise, remain on station and keep the peace."
+	missiondesc += "<BR> <B>1.</B> Serve the public trust."
+	missiondesc += "<BR> <B>2.</B> Protect the innocent."
+	missiondesc += "<BR> <B>3.</B> Uphold the law."
+	missiondesc += "<BR> <B>4.</B> Find the Undercover Cops."
+	missiondesc += "<BR> <B>5.</B> Detain Nanotrasen Security personnel if they harm any citizen."
+	missiondesc += "<BR> You can <B>see gangsters</B> using your <B>special sunglasses</B>."
 	to_chat(owner,missiondesc)
+	var/policy = get_policy(ROLE_FAMILIES)
+	if(policy)
+		to_chat(owner, policy)
+	var/mob/living/M = owner.current
+	M.playsound_local(M, 'sound/effects/families_police.ogg', 100, FALSE, pressure_affected = FALSE, use_reverb = FALSE)
 
 /datum/antagonist/ert/families/undercover_cop
 	name = "Undercover Cop"
 	role = "Undercover Cop"
 	outfit = /datum/outfit/families_police/beatcop
+	plasmaman_outfit = /datum/outfit/plasmaman/security
 	var/free_clothes = list(/obj/item/clothing/glasses/hud/spacecop/hidden,
 						/obj/item/clothing/under/rank/security/officer/beatcop,
 						/obj/item/clothing/head/spacepolice)
@@ -236,23 +296,25 @@
 	random_names = FALSE
 
 /datum/antagonist/ert/families/undercover_cop/on_gain()
-	for(var/C in free_clothes)
-		var/obj/O = new C(owner.current)
-		var/list/slots = list (
-			"backpack" = ITEM_SLOT_BACKPACK,
-			"left pocket" = ITEM_SLOT_LPOCKET,
-			"right pocket" = ITEM_SLOT_RPOCKET
-		)
-		var/mob/living/carbon/human/H = owner.current
-		var/equipped = H.equip_in_one_of_slots(O, slots)
-		if(!equipped)
-			to_chat(owner.current, "Unfortunately, you could not bring your [O] to this shift. You will need to find one.")
-			qdel(O)
+	if(istype(owner.current, /mob/living/carbon/human))
+		for(var/C in free_clothes)
+			var/obj/O = new C(owner.current)
+			var/list/slots = list (
+				"backpack" = ITEM_SLOT_BACKPACK,
+				"left pocket" = ITEM_SLOT_LPOCKET,
+				"right pocket" = ITEM_SLOT_RPOCKET
+			)
+			var/mob/living/carbon/human/H = owner.current
+			var/equipped = H.equip_in_one_of_slots(O, slots)
+			if(!equipped)
+				to_chat(owner.current, "Unfortunately, you could not bring your [O] to this shift. You will need to find one.")
+				qdel(O)
 	. = ..()
 
 
 /datum/antagonist/ert/families/undercover_cop/greet()
 	to_chat(owner, "<B><font size=3 color=red>You are the [name].</font></B>")
+	to_chat(owner, "<B><font size=3 color=red>You are NOT a Nanotrasen Employee. You work for the local government.</font></B>")
 
 	var/missiondesc = "You are an undercover police officer on board [station_name()]. You've been sent here by the Spinward Stellar Coalition because of suspected abusive behavior by the security department, and to keep tabs on a potential criminal organization operation."
 	missiondesc += "<BR><B>Your Mission</B>:"
